NodeDefsSelector.js: refactor to under webapp/components/survey and fix related
