+++ webapp/components/survey/NodeDefsSelector/components/NodeDefsSelector.scss
@@ -56,7 +56,7 @@
     padding: 0.3rem 0.5rem;
 
     .icon-left {
-      margin-left: 0px !important;
+      margin-left: 0 !important;
     }
 
     .icon {
