+++ webapp/components/survey/NodeDefsSelector/components/attributeSelector.js
