+++ webapp/components/survey/NodeDefsSelector/components/attributeSelector.scss
