+++ webapp/components/survey/NodeDefsSelector/components/attributesSelector.js
