+++ webapp/components/survey/NodeDefsSelector/components/attributesSelector.scss
