+++ webapp/components/survey/NodeDefsSelector/components/entitySelector.scss
