+++ webapp/components/survey/NodeDefsSelector/index.js
@@ -1,0 +1,3 @@
+export { default as NodeDefsSelectorView } from './NodeDefsSelector'
+
+export { default as NodeDefsSelectorAggregateView } from './NodeDefsSelectorAggregate'
