--- webapp/loggedin/surveyViews/nodeDefsSelector/index.js
@@ -1,3 +1,0 @@
-export { default as NodeDefsSelectorView } from './nodeDefsSelectorView'
-
-export { default as NodeDefsSelectorAggregateView } from './nodeDefsSelectorAggregateView'
